Move FilterOpHelpers from Frame -> Scene.
